requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
pytesseract==0.3.10
easyocr==1.7.0
opencv-python==4.8.1.78
numpy==1.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
loguru==0.7.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
requests==2.31.0